src/main/java/com/huo/service/impl/CategoryServiceImpl.java
package com.huo.service.impl;

import com.huo.entity.ProductCategory;
import com.huo.respository.ProductCategoryRespository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;
import com.huo.service.CategoryService;

import java.util.List;

/**
 * @Author: Huo
 * @Description:商品类目实现层
 * @Date: Create in 9:32 2020/4/23
 */
@Service
public class CategoryServiceImpl implements CategoryService {

    @Autowired
    ProductCategoryRespository respository;

    @Override
    public ProductCategory findOne(Integer categoryId) {
        return respository.findOne(categoryId);
    }

    @Override
    public List<ProductCategory> findAll() {
        return respository.findAll();
    }

    @Override
    public List<ProductCategory> findByCategoryTypeIn(List<Integer> categoryTypeList) {
        return respository.findByCategoryTypeIn(categoryTypeList);
    }

    @Override
    public ProductCategory save(ProductCategory productCategory) {
        return respository.save(productCategory);
    }
}
